terminado toda la api

+++ src/main/java/edu/unimagdalena/testcrud/repository/EstudianteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import edu.unimagdalena.testcrud.entity.Estudiante;
 
-public interface EstudianteRepository extends JpaRepository<Estudiante, Float> {
+public interface EstudianteRepository extends JpaRepository<Estudiante, Long> {
 	
 }
